Add arrays to hold books and persons objects
